stop external video if presenter changed
