+++ bigbluebutton-html5/imports/api/external-videos/server/handlers/stopExternalVideo.js
@@ -1,6 +1,5 @@
 import { check } from 'meteor/check';
 import Logger from '/imports/startup/server/logger';
-import Users from '/imports/api/users';
 import Meetings from '/imports/api/meetings';
 
 export default function handleStopExternalVideo({ header, body }, meetingId) {
@@ -9,10 +8,6 @@ export default function handleStopExternalVideo({ header, body }, meetingId) {
   check(meetingId, String);
   check(userId, String);
 
-  const user = Users.findOne({ meetingId: meetingId, userId: userId })
-
-  if (user && user.presenter) {
-      Logger.info(`User id=${userId} stop sharing an external video for meeting ${meetingId}`);
-      Meetings.update({ meetingId }, { $set: { externalVideoUrl: null } });
-  }
+  Logger.info(`External video stop sharing was initiated by:[${userId}] for meeting ${meetingId}`);
+  Meetings.update({ meetingId }, { $set: { externalVideoUrl: null } });
 }
