+++ bigbluebutton-html5/imports/api/external-videos/server/methods/stopWatchingExternalVideo.js
@@ -1,28 +1,21 @@
 import { check } from 'meteor/check';
 import Logger from '/imports/startup/server/logger';
 import Users from '/imports/api/users';
-import RedisPubSub from '/imports/startup/server/redis';
+import stopWatchingExternalVideoSystemCall from '/imports/api/external-videos/server/methods/stopWatchingExternalVideoSystemCall';
 import { extractCredentials } from '/imports/api/common/server/helpers';
 
-export default function stopWatchingExternalVideo(options) {
-  const REDIS_CONFIG = Meteor.settings.private.redis;
-  const CHANNEL = REDIS_CONFIG.channels.toAkkaApps;
-  const EVENT_NAME = 'StopExternalVideoPubMsg';
-
-  // when we call stopWatchingExternalVideo as a Meteor method from client side we obtain
-  // user credentials from the connection via extractCredentials
-  // However, we also call this function from server side (end meeting, user left, etc)
-  const { meetingId, requesterUserId } = this.userId ? extractCredentials(this.userId) : options;
+export default function stopWatchingExternalVideo() {
+  const { meetingId, requesterUserId } = extractCredentials(this.userId);
 
   try {
     check(meetingId, String);
     check(requesterUserId, String);
+
     const user = Users.findOne({ meetingId, userId: requesterUserId });
 
     if (user && user.presenter) {
-      const payload = { };
-      Logger.debug(`User id=${requesterUserId} sending ${EVENT_NAME} for meeting ${meetingId}`);
-      return RedisPubSub.publishUserMessage(CHANNEL, EVENT_NAME, meetingId, requesterUserId, payload);
+      // proceed and publish the event
+      stopWatchingExternalVideoSystemCall({ meetingId, requesterUserId });
     }
   } catch (error) {
     Logger.error(`Error on stop sharing an external video for meeting=${meetingId} ${error}`);
