+++ bigbluebutton-html5/imports/api/users/server/modifiers/changePresenter.js
@@ -1,7 +1,7 @@
 import Logger from '/imports/startup/server/logger';
 import Users from '/imports/api/users';
 import Meetings from '/imports/api/meetings';
-import stopWatchingExternalVideo from '/imports/api/external-videos/server/methods/stopWatchingExternalVideo';
+import stopWatchingExternalVideoSystemCall from '/imports/api/external-videos/server/methods/stopWatchingExternalVideoSystemCall';
 
 export default function changePresenter(presenter, userId, meetingId, changedBy) {
   const selector = {
@@ -19,7 +19,7 @@ export default function changePresenter(presenter, userId, meetingId, changedBy)
     const meeting = Meetings.findOne({ meetingId });
     if (meeting && meeting.externalVideoUrl) {
       Logger.info(`ChangePresenter:There is external video being shared. Stopping it due to presenter change, ${meeting.externalVideoUrl}`);
-      stopWatchingExternalVideo({ meetingId, requesterUserId: userId });
+      stopWatchingExternalVideoSystemCall({ meetingId, requesterUserId: 'system-presenter-changed' });
     }
 
     const numberAffected = Users.update(selector, modifier);
